+++ src/aux-server/aux-web/aux-player/scene/PlayerSimulation3D.ts
@@ -303,10 +303,23 @@ export class PlayerSimulation3D extends Simulation3D {
                 tags: { 'aux._userContext': this.context },
             });
 
+            // Send an event to newly entered file indicating that the given context was loaded.
+            await this.simulation.helper.action(
+                'onPlayerContextEnter',
+                [userFile],
+                {
+                    context: this.context,
+                }
+            );
+
             // Send an event to all files indicating that the given context was loaded.
-            await this.simulation.helper.action('onPlayerContextEnter', null, {
-                context: this.context,
-            });
+            await this.simulation.helper.action(
+                'onAnyPlayerContextEnter',
+                null,
+                {
+                    context: this.context,
+                }
+            );
 
             this._subs.push(
                 this.simulation.watcher